Move List into its own file
