+++ polygraph/types/tests/test_type_map.py
@@ -2,9 +2,10 @@ from unittest import TestCase
 
 from polygraph.types.decorators import field
 from polygraph.types.object_type import ObjectType
+from polygraph.types.list import List
 from polygraph.types.scalar import ID, Boolean, Float, Int, String
 from polygraph.types.schema import Schema
-from polygraph.types.type_builder import List, NonNull
+from polygraph.types.type_builder import NonNull
 from polygraph.types.union import Union
