+++ polygraph/types/tests/test_types.py
@@ -2,8 +2,9 @@ from unittest import TestCase, skip
 
 from polygraph.exceptions import PolygraphValueError
 from polygraph.types.api import typedef
+from polygraph.types.list import List
 from polygraph.types.scalar import Boolean, Int, String
-from polygraph.types.type_builder import List, NonNull
+from polygraph.types.type_builder import NonNull
 from polygraph.utils.trim_docstring import trim_docstring
